+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Restrict supported compression algorithms to `gzip`
+- Add support for (de-)compression of HTTP requests
 
 ## HCloud v2.1.0 (2024-07-28)
 
@@ -11,7 +11,6 @@
 - Add deprecation warning for `Pricing#traffic`
 - Add deprecation warning for `ServerType#included_traffic`
 - Fix crash on prices when using `Pricing` API
-- Add support for (de-)compression of HTTP requests
 
 ## HCloud v2.0.0 (2024-01-10)